+++ src/gtfsapp/id/GTFSID.java
@@ -1,27 +1,24 @@
 package gtfsapp.id;
 
-import java.util.ArrayList;
+import java.util.HashSet;
 
 /**
  * @author Michael Primeau
  * @version 1.0
  * @created 15-Apr-2020 1:20:18 PM
  */
-public abstract class GTFSID {
+public class GTFSID {
 
-    private static ArrayList<GTFSID> existingIDs = new ArrayList<>();
+    private static HashSet<GTFSID> existingIDs = new HashSet<>();
     private String id;
 
 
     /**
      * Constructor for GTFSID
-     * @param id String id to be associated with this object
+     * @param id String ID to be associated with this object
      */
     public GTFSID(String id) {
-        if(exists(id)) {
-            throw new IllegalArgumentException("id already exists");
-        }
-        this.id = id;
+        this.setID(id);
 
         // TODO
         // Not sure if this is allowed
@@ -29,19 +26,27 @@ public abstract class GTFSID {
     }
 
     /**
+     * Checks if ID argument already exists and sets GTFSID id to the passed in argument
+     * @param id String ID to set this objects ID to
+     * @return True if the ID was successfully set
+     */
+    public boolean setID(String id) {
+        for(GTFSID s:existingIDs) {
+            if(s.id.equals(id)) {
+                // Temporary Exception
+                throw new IllegalArgumentException("ID already exists");
+            }
+        }
+        this.id = id;
+        return true;
+    }
+
+    /**
      * Check if an ID already exists in the data structure
      * @param id ID to check for existence
      */
     public static boolean exists(GTFSID id) {
-
-        // TODO - michael, you can use ArrayList.contains(Object O) here :) -Grant
-
-        for(GTFSID gtfsid:existingIDs) {
-            if(gtfsid.equals(id)) {
-                return true;
-            }
-        }
-        return false;
+        return existingIDs.contains(id);
     }
 
     /**
@@ -49,12 +54,6 @@ public abstract class GTFSID {
      * @param id
      */
     public boolean equals(GTFSID id) {
-        /*
-        TODO:
-        We need an equals method for the exists() method to work.
-        We should compare both the ID and the type using ObjectOf()
-        This is because IDs of different types can have the same ID string
-        (e.g. StopID with ID string J12 != TripID with ID string J12)
-         */
+        return (this.id.equals(id.id) && (this.getClass() == id.getClass()));
     }
 }